// migrations/snapshots/b8f05b8e8a143aaedc49e98ed10fd54f7428de4130897f054fbabdc69a080d33/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly idCliente: CodecTypes['pg/int4@1']['output'];
      readonly nomeCliente: CodecTypes['sql/varchar@1']['output'] | null;
      readonly emailCliente: CodecTypes['sql/varchar@1']['output'] | null;
    };
    readonly TelCliente: {
      readonly idTelCliente: CodecTypes['pg/int4@1']['output'];
      readonly idCliente: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Teste: {
      readonly idTeste: CodecTypes['pg/int4@1']['output'];
      readonly nomeTeste: CodecTypes['sql/varchar@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly idCliente: CodecTypes['pg/int4@1']['input'];
      readonly nomeCliente: CodecTypes['sql/varchar@1']['input'] | null;
      readonly emailCliente: CodecTypes['sql/varchar@1']['input'] | null;
    };
    readonly TelCliente: {
      readonly idTelCliente: CodecTypes['pg/int4@1']['input'];
      readonly idCliente: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Teste: {
      readonly idTeste: CodecTypes['pg/int4@1']['input'];
      readonly nomeTeste: CodecTypes['sql/varchar@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly cliente: {
      readonly email_cliente: CodecTypes['sql/varchar@1']['output'] | null;
      readonly id_cliente: CodecTypes['pg/int4@1']['output'];
      readonly nome_cliente: CodecTypes['sql/varchar@1']['output'] | null;
    };
    readonly tel_cliente: {
      readonly id_cliente: CodecTypes['pg/int4@1']['output'] | null;
      readonly id_tel_cliente: CodecTypes['pg/int4@1']['output'];
    };
    readonly teste: {
      readonly id_Teste: CodecTypes['pg/int4@1']['output'];
      readonly nome_cliente: CodecTypes['sql/varchar@1']['output'] | null;
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly cliente: {
      readonly email_cliente: CodecTypes['sql/varchar@1']['input'] | null;
      readonly id_cliente: CodecTypes['pg/int4@1']['input'];
      readonly nome_cliente: CodecTypes['sql/varchar@1']['input'] | null;
    };
    readonly tel_cliente: {
      readonly id_cliente: CodecTypes['pg/int4@1']['input'] | null;
      readonly id_tel_cliente: CodecTypes['pg/int4@1']['input'];
    };
    readonly teste: {
      readonly id_Teste: CodecTypes['pg/int4@1']['input'];
      readonly nome_cliente: CodecTypes['sql/varchar@1']['input'] | null;
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;